fix: change githubId and related fields type from bigint to int for consistency

+++ src/repository/Account.ts
@@ -40,7 +40,7 @@ export class AccountRepo {
 	/**
 	 * ACCOUNT GITHUB that are the owner
 	 */
-	@Column({ type: 'bigint', nullable: true })
+	@Column({ type: 'int', nullable: true })
 	githubId?: number;
